Adapt layout
