+++ apps/desktop/src/app/base-layout/base-layout.component.html
@@ -1,0 +1,6 @@
+<nde-topbar></nde-topbar>
+<div
+  class="page-wrapper"
+  *ngIf="isLogged">
+  <router-outlet></router-outlet>
+</div>
