+++ apps/desktop/src/app/base-layout/base-layout.component.scss
@@ -1,0 +1,5 @@
+@import 'variables';
+
+.page-wrapper {
+  padding-top: var(--app-topbar-height);
+}
