+++ apps/desktop/src/app/base-layout/topbar.component.scss
@@ -1,4 +1,4 @@
-@import 'apps/dashboard/src/variables';
+@import 'variables';
 
 .top-bar {
   align-items: center;
