+++ libs/sync/src/lib/main-layout/main-layout.component.html
@@ -1,9 +1,4 @@
-<nsy-topbar></nsy-topbar>
 <nsy-sidebar></nsy-sidebar>
-<div class="page-wrapper">
-  <div
-    class="page-content"
-    *ngIf="isLogged">
-    <router-outlet></router-outlet>
-  </div>
+<div class="sync-container">
+  <router-outlet></router-outlet>
 </div>
