+++ libs/sync/src/lib/main-layout/main-layout.component.scss
@@ -1,13 +1,6 @@
 @import 'apps/dashboard/src/variables';
 
-.page-wrapper {
-  padding: 70px 0 0 rhythm(35);
-  height: 100%;
-  display: flex;
-  overflow: scroll;
-}
-
-.page-content {
+.sync-container {
+  margin-left: rhythm(35);
   padding: rhythm(5) rhythm(10);
-  width: 100%;
 }
