+++ libs/sync/src/lib/main-layout/sidebar.component.scss
@@ -5,7 +5,7 @@
   height: calc(100% - #{rhythm(9)});
   position: fixed;
   top: rhythm(9);
-  left: 0;
+  left: auto;
   overflow-y: auto;
   padding-top: rhythm(6);
   width: rhythm(35);
